backend/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
cohere>=4.9.0
qdrant-client==1.7.3
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.22.0
huggingface_hub>=0.20.0
transformers
torch
